+++ packages/at_client/lib/src/transformer/response_transformer/notification_response_transformer.dart
@@ -1,21 +1,20 @@
 import 'dart:async';
 
 import 'package:at_client/at_client.dart';
-import 'package:at_client/src/decryption_service/decryption.dart';
 import 'package:at_client/src/transformer/at_transformer.dart';
 import 'package:at_client/src/decryption_service/decryption_manager.dart';
-import 'package:meta/meta.dart';
 
 /// Class is responsible for decrypting the notification value/text-message data
 class NotificationResponseTransformer
     implements
         Transformer<Tuple<AtNotification, NotificationConfig>, AtNotification> {
   late final AtClient _atClient;
+  late final AtKeyDecryptionManager _decryptionManager;
 
-  @visibleForTesting
-  AtKeyDecryption? atKeyDecryption;
-
-  NotificationResponseTransformer(this._atClient);
+  NotificationResponseTransformer(this._atClient,
+      {AtKeyDecryptionManager? decrypterManager}) {
+    _decryptionManager = decrypterManager ?? AtKeyDecryptionManager(_atClient);
+  }
 
   @override
   Future<AtNotification> transform(
@@ -81,11 +80,9 @@ class NotificationResponseTransformer
   }
 
   Future<String> _getDecryptedValue(AtKey atKey, String? encryptedValue) async {
-    atKeyDecryption ??=
-        AtKeyDecryptionManager(_atClient).get(atKey, atKey.sharedWith!);
-    var decryptedValue =
-        await atKeyDecryption?.decrypt(atKey, encryptedValue?.trim());
+    final decrypter = _decryptionManager.get(atKey);
+    final decrypted = await decrypter.decrypt(atKey, encryptedValue?.trim());
     // Return decrypted value
-    return decryptedValue.toString().trim();
+    return decrypted.toString().trim();
   }
 }